func like and detail product

+++ bambu/database/migrations/2019_06_13_221917_create_articles_table.php
@@ -22,7 +22,7 @@ class CreateArticlesTable extends Migration
             $table->bigInteger('priceMajor');
             $table->bigInteger('priceTuB');
             $table->string('department');
-            $table->bigInteger('weight');
+            $table->decimal('weight');
             $table->string('photo'); //extension .jpg .png ..etc
             $table->boolean('gender');
